158 fixing upload image issue

+++ server/routes/courses.js
@@ -213,6 +213,9 @@ router.post('/myCoursesInfo', VerifyToken, async (req, res) => {
 
 router.post('/create', VerifyToken, async (req, res) => {
   try {
+
+    console.log('here = ' + req.body.urlImage)
+
     const course = new Course({
       name: req.body.name,
       description: req.body.description,
@@ -224,6 +227,8 @@ router.post('/create', VerifyToken, async (req, res) => {
     const savedCourse = await course.save();
 
     findCourse = await Course.findOne({ "name": req.body.name, "description": req.body.description }, '_id')
+
+    
     // console.log(findCourse._id)
     const updateUser = await User.updateOne(
       { _id: req.body.userID },
@@ -279,14 +284,20 @@ router.post('/removeEnrollment', VerifyToken, async (req, res) => {
 // TODO: Need to move this to fileMulter once I figure out why it's not getting called when it sits in fileMulter
 router.post('/removeFile', VerifyToken, async (req, res) => {
 
-  const pathname = req.body.imageName.split('/')
-  const imageName = pathname[pathname.length - 1]
-  console.log(imageName)
-  const path = 'public/' + req.body.courseID + '/' + imageName
-
   try {
-    fs.unlinkSync(path)
-    res.json({ 'status': 'module added' });
+
+    course = await Course.findOne({ _id: req.body.courseID }, 'urlImage')
+
+    if(course !== undefined)
+    {
+      const pathname = course.urlImage.split('/')
+      const imageName = pathname[pathname.length - 1]
+
+      const path = 'public/' + req.body.courseID + '/' + imageName
+
+      fs.unlinkSync(path)
+      res.json({ 'status': 'module added' });
+    }
 
   } catch (err) {
     console.error(err)
